[Desktop] Rename component user component folder comments functionalities without backend support
